+++ server/standalone/dbal/computeplan.go
@@ -3,6 +3,7 @@ package dbal
 import (
 	"errors"
 	"strconv"
+	"time"
 
 	sq "github.com/Masterminds/squirrel"
 	"github.com/jackc/pgx/v4"
@@ -10,50 +11,88 @@ import (
 	"github.com/owkin/orchestrator/lib/common"
 	orcerrors "github.com/owkin/orchestrator/lib/errors"
 	"github.com/owkin/orchestrator/lib/persistence"
+	"google.golang.org/protobuf/types/known/timestamppb"
 )
+
+type sqlComputePlan struct {
+	Key                      string
+	Owner                    string
+	DeleteIntermediaryModels bool
+	CreationDate             time.Time
+	Tag                      string
+	Metadata                 map[string]string
+	TaskCounts               persistence.ComputePlanTaskCount
+}
+
+// toRawComputePlan returns a compute plan without its computed properties.
+func (cp *sqlComputePlan) toRawComputePlan() *asset.ComputePlan {
+	return &asset.ComputePlan{
+		Key:                      cp.Key,
+		Owner:                    cp.Owner,
+		DeleteIntermediaryModels: cp.DeleteIntermediaryModels,
+		CreationDate:             timestamppb.New(cp.CreationDate),
+		Tag:                      cp.Tag,
+		Metadata:                 cp.Metadata,
+	}
+}
+
+// toComputePlan returns a compute plan with its computed properties.
+func (cp *sqlComputePlan) toComputePlan() *asset.ComputePlan {
+	plan := cp.toRawComputePlan()
+	plan.TaskCount = cp.TaskCounts.Total
+	plan.WaitingCount = cp.TaskCounts.Waiting
+	plan.TodoCount = cp.TaskCounts.Todo
+	plan.DoingCount = cp.TaskCounts.Doing
+	plan.CanceledCount = cp.TaskCounts.Canceled
+	plan.FailedCount = cp.TaskCounts.Failed
+	plan.DoneCount = cp.TaskCounts.Done
+	plan.Status = cp.TaskCounts.GetPlanStatus()
+	return plan
+}
 
 // ComputePlanExists returns true if a ComputePlan with the given key already exists
 func (d *DBAL) ComputePlanExists(key string) (bool, error) {
-	row := d.tx.QueryRow(d.ctx, `select count(id) from "compute_plans" where id=$1 and channel=$2`, key, d.channel)
+	stmt := getStatementBuilder().
+		Select("COUNT(key)").
+		From("compute_plans").
+		Where(sq.Eq{"key": key, "channel": d.channel})
+
+	row, err := d.queryRow(stmt)
+	if err != nil {
+		return false, err
+	}
 
 	var count int
-	err := row.Scan(&count)
+	err = row.Scan(&count)
 
 	return count == 1, err
 }
 
 // AddComputePlan stores a new ComputePlan in DB
 func (d *DBAL) AddComputePlan(plan *asset.ComputePlan) error {
-	stmt := `insert into "compute_plans" ("id", "asset", "channel") values ($1, $2, $3)`
-	_, err := d.tx.Exec(d.ctx, stmt, plan.GetKey(), plan, d.channel)
-	return err
+	stmt := getStatementBuilder().
+		Insert("compute_plans").
+		Columns("key", "channel", "owner", "delete_intermediary_models", "creation_date", "tag", "metadata").
+		Values(plan.Key, d.channel, plan.Owner, plan.DeleteIntermediaryModels, plan.CreationDate.AsTime(), plan.Tag, plan.Metadata)
+
+	return d.exec(stmt)
 }
 
 // GetComputePlan returns a ComputePlan by its key
 func (d *DBAL) GetComputePlan(key string) (*asset.ComputePlan, error) {
-	stmt := getStatementBuilder().Select().
-		Column("cp.asset").
-		Column(sq.Expr("COUNT(1)")).
-		Column(sq.Expr("COUNT(1) FILTER (WHERE t.status = 'STATUS_WAITING')")).
-		Column(sq.Expr("COUNT(1) FILTER (WHERE t.status = 'STATUS_TODO')")).
-		Column(sq.Expr("COUNT(1) FILTER (WHERE t.status = 'STATUS_DOING')")).
-		Column(sq.Expr("COUNT(1) FILTER (WHERE t.status = 'STATUS_CANCELED')")).
-		Column(sq.Expr("COUNT(1) FILTER (WHERE t.status = 'STATUS_FAILED')")).
-		Column(sq.Expr("COUNT(1) FILTER (WHERE t.status = 'STATUS_DONE')")).
-		From("compute_plans AS cp").
-		LeftJoin("compute_tasks AS t ON t.compute_plan_id = cp.id").
-		Where(sq.Eq{"cp.id": key}).
-		Where(sq.Eq{"cp.channel": d.channel}).
-		GroupBy("cp.id")
+	stmt := getStatementBuilder().
+		Select("key", "owner", "delete_intermediary_models", "creation_date", "tag", "metadata", "task_count", "waiting_count", "todo_count", "doing_count", "canceled_count", "failed_count", "done_count").
+		From("expanded_compute_plans").
+		Where(sq.Eq{"key": key, "channel": d.channel})
 
 	row, err := d.queryRow(stmt)
 	if err != nil {
 		return nil, err
 	}
 
-	plan := new(asset.ComputePlan)
-	var total, waiting, todo, doing, canceled, failed, done uint32
-	err = row.Scan(plan, &total, &waiting, &todo, &doing, &canceled, &failed, &done)
+	pl := new(sqlComputePlan)
+
+	err = row.Scan(&pl.Key, &pl.Owner, &pl.DeleteIntermediaryModels, &pl.CreationDate, &pl.Tag, &pl.Metadata, &pl.TaskCounts.Total, &pl.TaskCounts.Waiting, &pl.TaskCounts.Todo, &pl.TaskCounts.Doing, &pl.TaskCounts.Canceled, &pl.TaskCounts.Failed, &pl.TaskCounts.Done)
 	if err != nil {
 		if errors.Is(err, pgx.ErrNoRows) {
 			return nil, orcerrors.NewNotFound("computeplan", key)
@@ -61,26 +100,23 @@ func (d *DBAL) GetComputePlan(key string) (*asset.ComputePlan, error) {
 		return nil, err
 	}
 
-	plan.TaskCount = total
-	plan.WaitingCount = waiting
-	plan.TodoCount = todo
-	plan.DoingCount = doing
-	plan.CanceledCount = canceled
-	plan.FailedCount = failed
-	plan.DoneCount = done
-	plan.Status = getPlanStatus(total, done, doing, waiting, failed, canceled)
-
-	return plan, nil
+	return pl.toComputePlan(), nil
 }
 
 // GetRawComputePlan returns a compute plan without its computed properties.
 func (d *DBAL) GetRawComputePlan(key string) (*asset.ComputePlan, error) {
-	query := `select asset from compute_plans where id=$1 and channel=$2`
+	stmt := getStatementBuilder().
+		Select("key", "owner", "delete_intermediary_models", "creation_date", "tag", "metadata").
+		From("compute_plans").
+		Where(sq.Eq{"key": key, "channel": d.channel})
 
-	row := d.tx.QueryRow(d.ctx, query, key, d.channel)
+	row, err := d.queryRow(stmt)
+	if err != nil {
+		return nil, err
+	}
 
-	plan := new(asset.ComputePlan)
-	err := row.Scan(plan)
+	pl := new(sqlComputePlan)
+	err = row.Scan(&pl.Key, &pl.Owner, &pl.DeleteIntermediaryModels, &pl.CreationDate, &pl.Tag, &pl.Metadata)
 	if err != nil {
 		if errors.Is(err, pgx.ErrNoRows) {
 			return nil, orcerrors.NewNotFound("computeplan", key)
@@ -88,7 +124,7 @@ func (d *DBAL) GetRawComputePlan(key string) (*asset.ComputePlan, error) {
 		return nil, err
 	}
 
-	return plan, nil
+	return pl.toRawComputePlan(), nil
 }
 
 func (d *DBAL) QueryComputePlans(p *common.Pagination, filter *asset.PlanQueryFilter) ([]*asset.ComputePlan, common.PaginationToken, error) {
@@ -97,26 +133,17 @@ func (d *DBAL) QueryComputePlans(p *common.Pagination, filter *asset.PlanQueryFi
 		return nil, "", err
 	}
 
-	stmt := getStatementBuilder().Select().
-		Column("cp.asset").
-		Column(sq.Expr("COUNT(1)")).
-		Column(sq.Expr("COUNT(1) FILTER (WHERE t.status = 'STATUS_WAITING')")).
-		Column(sq.Expr("COUNT(1) FILTER (WHERE t.status = 'STATUS_TODO')")).
-		Column(sq.Expr("COUNT(1) FILTER (WHERE t.status = 'STATUS_DOING')")).
-		Column(sq.Expr("COUNT(1) FILTER (WHERE t.status = 'STATUS_CANCELED')")).
-		Column(sq.Expr("COUNT(1) FILTER (WHERE t.status = 'STATUS_FAILED')")).
-		Column(sq.Expr("COUNT(1) FILTER (WHERE t.status = 'STATUS_DONE')")).
-		From("compute_plans AS cp").
-		LeftJoin("compute_tasks AS t ON t.compute_plan_id = cp.id").
-		Where(sq.Eq{"cp.channel": d.channel}).
-		GroupBy("cp.id").
-		OrderBy("cp.asset->>'creationDate' ASC", "cp.id ASC").
+	stmt := getStatementBuilder().
+		Select("key", "owner", "delete_intermediary_models", "creation_date", "tag", "metadata", "task_count", "waiting_count", "todo_count", "doing_count", "canceled_count", "failed_count", "done_count").
+		From("expanded_compute_plans").
+		Where(sq.Eq{"channel": d.channel}).
+		OrderBy("creation_date ASC, key ASC").
 		Offset(uint64(offset)).
 		// Fetch page size + 1 elements to determine whether there is a next page
 		Limit(uint64(p.Size + 1))
 
 	if filter.Owner != "" {
-		stmt = stmt.Where(sq.Eq{"asset->>'owner'": filter.Owner})
+		stmt = stmt.Where(sq.Eq{"owner": filter.Owner})
 	}
 
 	rows, err := d.query(stmt)
@@ -129,25 +156,14 @@ func (d *DBAL) QueryComputePlans(p *common.Pagination, filter *asset.PlanQueryFi
 	var count int
 
 	for rows.Next() {
-		plan := new(asset.ComputePlan)
+		pl := new(sqlComputePlan)
 
-		var total, waiting, todo, doing, canceled, failed, done uint32
-
-		err = rows.Scan(plan, &total, &waiting, &todo, &doing, &canceled, &failed, &done)
+		err = rows.Scan(&pl.Key, &pl.Owner, &pl.DeleteIntermediaryModels, &pl.CreationDate, &pl.Tag, &pl.Metadata, &pl.TaskCounts.Total, &pl.TaskCounts.Waiting, &pl.TaskCounts.Todo, &pl.TaskCounts.Doing, &pl.TaskCounts.Canceled, &pl.TaskCounts.Failed, &pl.TaskCounts.Done)
 		if err != nil {
 			return nil, "", err
 		}
 
-		plan.TaskCount = total
-		plan.WaitingCount = waiting
-		plan.TodoCount = todo
-		plan.DoingCount = doing
-		plan.CanceledCount = canceled
-		plan.FailedCount = failed
-		plan.DoneCount = done
-		plan.Status = getPlanStatus(total, done, doing, waiting, failed, canceled)
-
-		plans = append(plans, plan)
+		plans = append(plans, pl.toComputePlan())
 		count++
 
 		if count == int(p.Size) {
@@ -165,18 +181,4 @@ func (d *DBAL) QueryComputePlans(p *common.Pagination, filter *asset.PlanQueryFi
 	}
 
 	return plans, bookmark, nil
-}
-
-// getPlanStatus determines ComputePlan status from its task counts
-func getPlanStatus(total, done, doing, waiting, failed, canceled uint32) asset.ComputePlanStatus {
-	count := persistence.ComputePlanTaskCount{
-		Total:    (int)(total),
-		Waiting:  (int)(waiting),
-		Doing:    (int)(doing),
-		Canceled: (int)(canceled),
-		Failed:   (int)(failed),
-		Done:     (int)(done),
-	}
-
-	return count.GetPlanStatus()
 }