+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Remove compute plan asset column (#618)
+
 ## [0.7.0] - 2022-03-29
 
 ### Added